Consolidate processor loading
